+++ src/util/compute.js
